+++ lib/utils/log.utils.ts
@@ -1,0 +1,5 @@
+
+export const TerminalLog = (message: string): void => {
+    process.stdout.write.bind(process.stdout)(message);
+}
+export default TerminalLog;