+++ CHANGELOG.md
@@ -6,6 +6,26 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+### [0.0.1-beta.2] - 2022-01-29
+
+### Added
+
+- log: format messages for each category
+- [ ERROR ]: Step.error(props)
+- [ DEBUG ]: Step.debug(props)
+- [ INFO  ]: Step.info(props)
+- [ FATAL ]: Step.fatal(props)
+- [ WARN  ]: Step.warn(props)
+- [ STACK ]: Step.stack(props)
+- [ UNDEF ]: When create step with invalid type
+- [ LOG   ]: When create a global log - Log.ini()
+
+### Change
+
+- log: remove console.log and use process.stdout
+
+---
+
 ### [0.0.1-beta.1] - 2022-01-29
 
 ### Added